update vagrant file to use install script
try adding env var for roles to be used in matrix

fix git task

move functions to install.sh with optional playbook run.
Removed redundant init.sh

+++ scripts/install.sh
@@ -1,10 +1,47 @@
 #!/bin/bash
 USERNAME=$(whoami)
-if [[ -z $1 ]]; then
+if [[ $1 == "play" ]]; then
+    RUN_PLAYBOOK="play"
+fi
+
+if [[ -z $2 ]]; then
     TAGS=""
 else
     TAGS="--tags $1"
 fi
+
+function check_user() {
+     user=${1:-"$(whoami)"}
+     echo "USER: ${user}"
+     if [[ $user ==  'root' ]]; then
+          echo -e "ERROR, user is root!\nRe-run script without sudo!"
+          exit 1
+     fi
+}
+
+function apt_dependencies() {
+     sudo apt-get update
+     sudo apt-get upgrade -y
+     sudo apt-get install -y python3-pip
+}
+
+function pip_dependencies() {
+	if ! grep -Fxq ".local/bin" ~/.profile ; then 
+          echo "PATH=$PATH:/home/${USERNAME}/.local/bin" >> ~/.profile
+     fi
+     pip3 install setuptools wheel
+     pip3 install virtualenv
+     which virtualenv
+}
+
+function installer_virtualenv() {
+     mkdir -p /home/${user}/.installer
+     virtualenv /home/${user}/.installer
+     source /home/${user}/.installer/bin/activate
+     pip3 install ansible==2.10.3
+     ansible --version
+}
+
 set -e
 if [[ -f /home/${USERNAME}/.local/bin/virtualenv ]]; then
     echo "Virtualenv detected. Skipping init script setup."
@@ -17,12 +54,10 @@ if [[ -f /home/${USERNAME}/.local/bin/virtualenv ]]; then
     else
         echo "No virtualenv detected for running ansible playbook."
         echo "Setting up installer virtualenv..."
-    	source ./scripts/init.sh 
         installer_virtualenv
 	    . /home/${USERNAME}/.installer/bin/activate
 	fi
 else
-	source ./scripts/init.sh 
     check_user ${USERNAME}
     apt_dependencies
     pip_dependencies
@@ -30,11 +65,14 @@ else
 	. /home/${USERNAME}/.installer/bin/activate
 fi
 
-sudo -k
-ansible-playbook \
-    --user ${USERNAME} \
-    --inventory inventory/hosts.yaml \
-    ${TAGS} \
-    --ask-become-pass \
-    playbooks/machine_setup.yaml
-rm -rf ~/.installer
+if [[ -z $RUN_PLAYBOOK ]]; then
+
+    sudo -k
+    ansible-playbook \
+        --user ${USERNAME} \
+        --inventory inventory/hosts.yaml \
+        ${TAGS} \
+        --ask-become-pass \
+        playbooks/machine_setup.yaml
+    rm -rf ~/.installer
+fi
